move provider_authorize helpers into scopes.ts
